requirements.txt
# Low Latency Trading Platform - Dependencies

# Core Dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Machine Learning Framework
tensorflow>=2.10.0
keras>=2.10.0

# Trading APIs
alpaca-trade-api>=2.3.0
ccxt>=3.0.0  # Multi-exchange support

# High Performance & Low Latency
asyncio
aiohttp>=3.8.0
# uvloop>=0.17.0  # Faster event loop
cython>=0.29.0  # For performance critical code
numba>=0.56.0  # JIT compilation for numpy

# Message Queue & Streaming
redis>=4.3.0
kafka-python>=2.0.0
websockets>=10.0

# Database & Storage
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymongo>=4.0.0  # MongoDB for tick data

# Risk Management & Monitoring
psutil>=5.8.0
prometheus-client>=0.14.0
structlog>=22.1.0

# Configuration & Environment
python-dotenv>=0.19.0
pydantic>=1.10.0
click>=8.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-mock>=3.6.0
pytest-benchmark>=3.4.0

# Networking & Security
cryptography>=37.0.0
requests>=2.28.0

# Data Processing & Analysis
scipy>=1.9.0
matplotlib>=3.5.0
plotly>=5.10.0

# Performance Monitoring
py-spy>=0.3.0
memory-profiler>=0.60.0